@@ -16,21 +16,19 @@ classifiers = [
     "Intended Audience :: Developers",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Operating System :: OS Independent",
 ]
-license = { text = "MIT" }
 dynamic = ["version"]
 readme = "README.md"
 dependencies = [
     "higlass-schema>=0.2.0",
     "anywidget>=0.9.0",
-    "typing-extensions ; python_version<'3.9'",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 urls = { homepage = "https://github.com/higlass/higlass-python" }
 
 [dependency-groups]
@@ -49,10 +47,6 @@ packages = ["src/higlass"]
 
 [tool.hatch.version]
 source = "vcs"
-
-[tool.ruff]
-line-length = 88
-target-version = "py38"
 
 [tool.ruff.lint]
 extend-select = [
