@@ -1,12 +1,12 @@
 from __future__ import annotations
 
 import uuid
-from typing import Literal, TypeVar, Union
+from typing import Literal, TypeVar
 
 import higlass_schema as hgs
 from pydantic import BaseModel
 
-TrackType = Union[hgs.EnumTrackType, Literal["heatmap"]]
+TrackType = hgs.EnumTrackType | Literal["heatmap"]
 TrackPosition = Literal["center", "top", "left", "bottom", "center", "whole", "gallery"]
 
 track_default_position: dict[str, TrackPosition] = {
